requirements.txt
torch>=2.1.0
torchvision>=0.16.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
PyYAML>=6.0
pydantic>=2.6.0
tqdm>=4.66.0
tensorboard>=2.15.0
onnx>=1.16.0
onnxruntime>=1.17.0

